+++ test/SpeedJumpIrmTest.sol
@@ -19,14 +19,13 @@ contract AdaptativeCurveIrmTest is Test {
     int256 internal constant CURVE_STEEPNESS = 4 ether;
     int256 internal constant ADJUSTMENT_SPEED = int256(50 ether) / 365 days;
     int256 internal constant TARGET_UTILIZATION = 0.9 ether;
-    uint256 internal constant INITIAL_RATE_AT_TARGET = uint256(0.01 ether) / 365 days;
+    int256 internal constant INITIAL_RATE_AT_TARGET = int256(0.01 ether) / 365 days;
 
     AdaptativeCurveIrm internal irm;
     MarketParams internal marketParams = MarketParams(address(0), address(0), address(0), address(0), 0);
 
     function setUp() public {
-        irm =
-        new AdaptativeCurveIrm(address(this), CURVE_STEEPNESS, ADJUSTMENT_SPEED, TARGET_UTILIZATION, int256(INITIAL_RATE_AT_TARGET));
+        irm = new AdaptativeCurveIrm(address(this), CURVE_STEEPNESS, ADJUSTMENT_SPEED, TARGET_UTILIZATION, INITIAL_RATE_AT_TARGET);
         vm.warp(90 days);
     }
 
@@ -39,7 +38,7 @@ contract AdaptativeCurveIrmTest is Test {
         Market memory market;
 
         assertApproxEqRel(
-            irm.borrowRate(marketParams, market), INITIAL_RATE_AT_TARGET / 4, 0.0001 ether, "avgBorrowRate"
+            irm.borrowRate(marketParams, market), uint256(INITIAL_RATE_AT_TARGET / 4), 0.0001 ether, "avgBorrowRate"
         );
         assertEq(irm.rateAtTarget(marketParams.id()), INITIAL_RATE_AT_TARGET, "rateAtTarget");
     }
@@ -49,7 +48,7 @@ contract AdaptativeCurveIrmTest is Test {
         market.totalBorrowAssets = 1 ether;
         market.totalSupplyAssets = 1 ether;
 
-        assertEq(irm.borrowRate(marketParams, market), INITIAL_RATE_AT_TARGET * 4, "avgBorrowRate");
+        assertEq(irm.borrowRate(marketParams, market), uint256(INITIAL_RATE_AT_TARGET * 4), "avgBorrowRate");
         assertEq(irm.rateAtTarget(marketParams.id()), INITIAL_RATE_AT_TARGET, "rateAtTarget");
     }
 
@@ -58,7 +57,7 @@ contract AdaptativeCurveIrmTest is Test {
         market.totalBorrowAssets = 0.9 ether;
         market.totalSupplyAssets = 1 ether;
 
-        assertEq(irm.borrowRate(marketParams, market), INITIAL_RATE_AT_TARGET, "avgBorrowRate");
+        assertEq(irm.borrowRate(marketParams, market), uint256(INITIAL_RATE_AT_TARGET), "avgBorrowRate");
         assertEq(irm.rateAtTarget(marketParams.id()), INITIAL_RATE_AT_TARGET, "rateAtTarget");
     }
 
@@ -82,7 +81,7 @@ contract AdaptativeCurveIrmTest is Test {
         vm.assume(market.totalSupplyAssets >= market.totalBorrowAssets);
 
         uint256 avgBorrowRate = irm.borrowRate(marketParams, market);
-        uint256 rateAtTarget = irm.rateAtTarget(marketParams.id());
+        int256 rateAtTarget = irm.rateAtTarget(marketParams.id());
 
         assertEq(avgBorrowRate, _curve(int256(INITIAL_RATE_AT_TARGET), _err(market)), "avgBorrowRate");
         assertEq(rateAtTarget, INITIAL_RATE_AT_TARGET, "rateAtTarget");
@@ -96,7 +95,7 @@ contract AdaptativeCurveIrmTest is Test {
         emit BorrowRateUpdate(
             marketParams.id(),
             _curve(int256(INITIAL_RATE_AT_TARGET), _err(market)),
-            _expectedRateAtTarget(marketParams.id(), market)
+            uint256(_expectedRateAtTarget(marketParams.id(), market))
         );
         irm.borrowRate(marketParams, market);
     }
@@ -106,7 +105,7 @@ contract AdaptativeCurveIrmTest is Test {
         vm.assume(market.totalSupplyAssets >= market.totalBorrowAssets);
 
         uint256 avgBorrowRate = irm.borrowRateView(marketParams, market);
-        uint256 rateAtTarget = irm.rateAtTarget(marketParams.id());
+        int256 rateAtTarget = irm.rateAtTarget(marketParams.id());
 
         assertEq(avgBorrowRate, _curve(int256(INITIAL_RATE_AT_TARGET), _err(market)), "avgBorrowRate");
         assertEq(rateAtTarget, 0, "prevBorrowRate");
@@ -121,7 +120,7 @@ contract AdaptativeCurveIrmTest is Test {
         vm.assume(market1.totalSupplyAssets >= market1.totalBorrowAssets);
         market1.lastUpdate = uint128(bound(market1.lastUpdate, 0, block.timestamp - 1));
 
-        uint256 expectedRateAtTarget = _expectedRateAtTarget(marketParams.id(), market1);
+        int256 expectedRateAtTarget = _expectedRateAtTarget(marketParams.id(), market1);
         uint256 expectedAvgRate = _expectedAvgRate(marketParams.id(), market1);
 
         uint256 borrowRateView = irm.borrowRateView(marketParams, market1);
@@ -141,7 +140,7 @@ contract AdaptativeCurveIrmTest is Test {
         vm.assume(market1.totalSupplyAssets >= market1.totalBorrowAssets);
         market1.lastUpdate = uint128(block.timestamp);
 
-        uint256 expectedRateAtTarget = _expectedRateAtTarget(marketParams.id(), market1);
+        int256 expectedRateAtTarget = _expectedRateAtTarget(marketParams.id(), market1);
         uint256 expectedAvgRate = _expectedAvgRate(marketParams.id(), market1);
 
         uint256 borrowRateView = irm.borrowRateView(marketParams, market1);
@@ -161,7 +160,7 @@ contract AdaptativeCurveIrmTest is Test {
         market1.totalSupplyAssets = market0.totalSupplyAssets;
         market1.lastUpdate = uint128(bound(market1.lastUpdate, 0, block.timestamp - 1));
 
-        uint256 expectedRateAtTarget = _expectedRateAtTarget(marketParams.id(), market1);
+        int256 expectedRateAtTarget = _expectedRateAtTarget(marketParams.id(), market1);
         uint256 expectedAvgRate = _expectedAvgRate(marketParams.id(), market1);
 
         uint256 borrowRateView = irm.borrowRateView(marketParams, market1);
@@ -190,14 +189,14 @@ contract AdaptativeCurveIrmTest is Test {
         );
     }
 
-    function _expectedRateAtTarget(Id id, Market memory market) internal view returns (uint256) {
+    function _expectedRateAtTarget(Id id, Market memory market) internal view returns (int256) {
         int256 rateAtTarget = int256(irm.rateAtTarget(id));
         int256 speed = ADJUSTMENT_SPEED.wMulDown(_err(market));
         uint256 elapsed = (rateAtTarget > 0) ? block.timestamp - market.lastUpdate : 0;
         int256 linearAdaptation = speed * int256(elapsed);
         int256 adaptationMultiplier = MathLib.wExp(linearAdaptation);
         return (rateAtTarget > 0)
-            ? uint256(rateAtTarget.wMulDown(adaptationMultiplier).bound(irm.MIN_RATE_AT_TARGET(), irm.MAX_RATE_AT_TARGET()))
+            ? rateAtTarget.wMulDown(adaptationMultiplier).bound(irm.MIN_RATE_AT_TARGET(), irm.MAX_RATE_AT_TARGET())
             : INITIAL_RATE_AT_TARGET;
     }
